password validation

+++ src/components/userInfo/PasswordChange.js
@@ -1,9 +1,13 @@
 /* jshint esversion: 8 */
 import { React, useState } from "react";
+import PasswordChecklist from "react-password-checklist";
 
 const PasswordChange = ({ user }) => {
-  const [oldPassword, setOldPassword] = useState();
-  const [newPassword, setNewPassword] = useState();
+  const [oldPassword, setOldPassword] = useState("");
+  const [newPassword, setNewPassword] = useState("");
+
+  const [passwordAgain, setPasswordAgain] = useState("");
+  const [isValidPass, setIsValidPass] = useState(false);
 
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
@@ -11,36 +15,44 @@ const PasswordChange = ({ user }) => {
   const userId = JSON.parse(localStorage.getItem("user"))._id;
   const token = localStorage.getItem("token");
 
+  const validPass = (isIt) => {
+    setIsValidPass(isIt);
+  };
+
   const changePass = async (e) => {
     e.preventDefault();
-    const userPasses = { oldPassword, newPassword };
-    try {
-      const response = await fetch(
-        `https://sp-backend-b70z.onrender.com/api/v1/users/changePassword/${userId}`,
-        {
-          method: "PATCH",
-          headers: {
-            "Content-type": "application/json",
-            "Authorization": `Bearer ${token}`,
-          },
-          body: JSON.stringify(userPasses),
-        }
-      );
-      if (!response.ok) {
-        throw new Error(
-          `This is an HTTP error: The status is ${response.status}`
+    if (isValidPass) {
+      const userPasses = { oldPassword, newPassword };
+      try {
+        const response = await fetch(
+          `https://sp-backend-b70z.onrender.com/api/v1/users/changePassword/${userId}`,
+          {
+            method: "PATCH",
+            headers: {
+              "Content-type": "application/json",
+              "Authorization": `Bearer ${token}`,
+            },
+            body: JSON.stringify(userPasses),
+          }
         );
+        if (!response.ok) {
+          throw new Error(
+            `This is an HTTP error: The status is ${response.status}`
+          );
+        }
+        // let userData = await response.json();
+        // setUserCopy(userData);
+        // setError(null);
+      } catch (err) {
+        setError(err.message);
+        // setUser(null);
+      } finally {
+        setLoading(false);
+        localStorage.clear();
+        window.location.reload();
       }
-      // let userData = await response.json();
-      // setUserCopy(userData);
-      // setError(null);
-    } catch (err) {
-      setError(err.message);
-      // setUser(null);
-    } finally {
-      setLoading(false);
-      localStorage.clear();
-      window.location.reload();
+    } else {
+      console.log(isValidPass);
     }
   };
 
@@ -62,19 +74,39 @@ const PasswordChange = ({ user }) => {
           <input
             type="password"
             value={newPassword}
+            id="newPassword"
             onChange={(e) => setNewPassword(e.target.value)}
             placeholder="Enter a password"
           />
         </div>
-        {/* <div className="info-field">
-          <p className="a-field">Re-enter New Password</p>
+        <div className="info-field">
+          <p className="a-field">Confirm Password</p>
           <input
+            value={passwordAgain}
+            onChange={(e) => setPasswordAgain(e.target.value)}
             type="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            placeholder="Enter a password"
+            id="passwordAgain"
+            placeholder="password"
           />
-        </div> */}
+        </div>
+        <div id="passError">
+          <PasswordChecklist
+            rules={[
+              "minLength",
+              "specialChar",
+              "number",
+              "capital",
+              "match",
+              "lowercase",
+            ]}
+            minLength={5}
+            value={newPassword}
+            valueAgain={passwordAgain}
+            onChange={(isValid) => {
+              validPass(isValid);
+            }}
+          />
+        </div>
         <button>Save Password</button>
       </form>
     </div>
